Move inhabited check to subs module
